Include Application JSON in the template using a helper
Because we want a single source of truth for our JSON API examples,
provide a way for the template to get at a canned response. We can
transform this as appropriate for other API responses.

Middleman supports custom Ruby functions in templates via helpers
(https://middlemanapp.com/basics/helper-methods/#custom-defined-helpers)

We’ve created a module with the functions we need so we can spec it
properly.