Add Capybara and configure it to run Middleman's server
Following https://github.com/SimonRice/middleman-rspec/blob/master/spec/spec_helper.rb,
configure Capybara to use the Middleman app as its testing app.

Because the govuk_tech_docs gem automatically includes LiveReload, and
LiveReload runs a separate server process, running the specs when a dev
server is already running causes the livereload server to crash with an
exception.

To get around this we define a new middleman environment
(https://middlemanapp.com/advanced/configuration/#environment-specific-settings)
called 'test' which disables livereload. The middleman docs say it's
possible to define environments using blocks in config.rb, but this didn't work
for me, perhaps because of something going on in the govuk_tech_docs
gem. (I could not get livereload to run when the call to
GovukTechDocs.configure was wrapped in a block).