Hyphenate rather than underscore anchor links
Would have preferred to cover this an integration spec, but
`current_url` in a Capybara context can't get us an anchor link (as
far as I can see).

This has ended up a little brittle, but I err on the side of having
the test until it gets annoying.

This commit also eliminates the line breaks in the generated HTML for
the ease of testing - I couldn't find evidence they were necessary.