Add HESA ITT details to the Application entity
This information

- is optional
- is only available once the student is enrolled
- is encoded according to the published HESA encoding for this cycle

If the application is not in the enrolled state this key is null in the
top-level application entity.

If the application is in the enrolled state this key is present. Its
children may have null values if the candidate declined to provide the
information, or an explicit "Information refused" code if HESA supports
it on that entity.

 I think this is the best we can do for now. We need to establish with
 HESA whether we can support "information refused" on sex and disability
 too - there's some uncertainty in the docs.