Copy the OpenAPI renderer into this project
There are some bugs in the functionality to render the OpenAPI
specification. We'd like to fix the bugs in the tech docs gem itself
(https://github.com/alphagov/tech-docs-gem), but Pull Requests to that
project take a while to get reviewed. In the meantime, we can iterate
on the template inside this application by copy-pasting the templates.
Once we're happy with our changes we can push the changes upstream.